created the component result that is rerender everytime the results are updated,
in the result component, if we have results then it returns a div with the word that was inputed and we go through each meaning('cause some words have multiple meanings) and within each of them we send the meaning to the Meaning component. In the Meaning component, we inclued  and dsiplayed partOfSpeech and loop through all the definitions and display the definition and the example.